refactor(collection): moved helper functions to utils
